requirements.txt
clint
requests
PyYaml